requirements.txt
asgiref==3.8.1
boto3==1.35.0
botocore==1.35.0
dj-database-url==2.2.0
Django==4.2.15
django-cors-headers==4.4.0
django-ses==4.1.1
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
whitenoise==6.7.0